Revise skipped tests, round 2 (#134)
* Remove Atlantis ry op. test

* Relocate test_conditional_operation

* Adjust QObj test name and desc.

* Rename file and test class

* Revamp setUp, test all backends

* Improve docstrings/naming

* Correct simulator -> real device

* Tweak docstring
